Refactor pull and progressacitivity
