+++ app/src/main/java/org/eyeseetea/malariacare/LoginActivity.java
@@ -197,7 +197,7 @@ public class LoginActivity extends AbsLoginActivity {
 
         //Populate locally
         try {
-            PullController.getInstance().wipeDatabase();
+            PopulateDB.wipeDatabase();
             PopulateDB.populateDB(getAssets());
         } catch (Exception ex) {
         }
