+++ app/src/main/java/org/eyeseetea/malariacare/ProgressActivity.java
@@ -35,17 +35,14 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import org.eyeseetea.malariacare.data.database.iomodules.dhis.exporter.PushController;
-import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.SyncProgressStatus;
 import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.models.EventExtended;
 import org.eyeseetea.malariacare.data.database.model.Survey;
-import org.eyeseetea.malariacare.data.database.utils.PopulateDB;
 import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
-import org.eyeseetea.malariacare.data.remote.SdkController;
 import org.eyeseetea.malariacare.data.repositories.UserAccountRepository;
-import org.eyeseetea.malariacare.domain.entity.PullFilters;
-import org.eyeseetea.malariacare.domain.entity.PullStep;
 import org.eyeseetea.malariacare.domain.usecase.LogoutUseCase;
-import org.eyeseetea.malariacare.domain.usecase.PullUseCase;
+import org.eyeseetea.malariacare.domain.usecase.pull.PullFilters;
+import org.eyeseetea.malariacare.domain.usecase.pull.PullStep;
+import org.eyeseetea.malariacare.domain.usecase.pull.PullUseCase;
 import org.eyeseetea.malariacare.layout.dashboard.builder.AppSettingsBuilder;
 import org.eyeseetea.malariacare.utils.Constants;
 
@@ -55,16 +52,11 @@ import java.util.List;
 
 public class ProgressActivity extends Activity {
 
-    private static final String TAG = ".ProgressActivity";
-
-
     public static final int NUMBER_OF_MONTHS = 6;
-
     /**
      * Intent param that tells what to do (push, pull or push before pull)
      */
     public static final String TYPE_OF_ACTION = "TYPE_OF_ACTION";
-
     /**
      * Intent param that tells what do before push
      */
@@ -73,17 +65,14 @@ public class ProgressActivity extends Activity {
      * To pull data from server
      */
     public static final int ACTION_PULL = 0;
-
     /**
      * To push a single survey to server
      */
     public static final int ACTION_PUSH = 1;
-
     /**
      * To dont show the survey pushed feedback
      */
     public static final int DONT_SHOW_FEEDBACK = 1;
-
     /**
      * To show the survey pushed feedback
      */
@@ -92,7 +81,7 @@ public class ProgressActivity extends Activity {
      * To push every unsent data to server before pulling metadata
      */
     public static final int ACTION_PUSH_BEFORE_PULL = 2;
-
+    private static final String TAG = ".ProgressActivity";
     /**
      * Num of expected steps while pulling
      */
@@ -121,19 +110,49 @@ public class ProgressActivity extends Activity {
      * Reference to progress message
      */
     public static TextView textView;
-    /**
-     * Reference required for testing purposes
-     */
-    AlertDialog alertDialog;
     static boolean pullAfterPushInProgress;
-    static Handler handler;
-    public static Activity progressActivity;
     static boolean isOnPause = true;
-
     //Check intent params
     static Intent intent;
-
     public PullUseCase mPullUseCase = new PullUseCase();
+
+    /**
+     * Prints the step in the progress bar
+     */
+    public static void step(final String msg) {
+        final int currentProgress = progressBar.getProgress();
+        progressBar.setProgress(currentProgress + 1);
+        textView.setText(msg);
+    }
+
+    private static void annotateFirstPull(boolean value) {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(
+                PreferencesState.getInstance().getContext());
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        editor.putBoolean(
+                PreferencesState.getInstance().getContext().getString(R.string.pull_metadata),
+                value);
+        editor.commit();
+    }
+
+    /**
+     * Tells if a push is required
+     */
+    private static boolean isAPush() {
+        //A push before pull
+        if (pullAfterPushInProgress) {
+            return false;
+        }
+
+        //Not a pull -> is a Push
+        return (intent != null && intent.getIntExtra(TYPE_OF_ACTION, ACTION_PULL) != ACTION_PULL);
+    }
+
+    private static String getDialogTitle(boolean isAPush) {
+        int stringId =
+                isAPush ? R.string.dialog_title_push_response : R.string.dialog_title_pull_response;
+        return PreferencesState.getInstance().getContext().getString(stringId);
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -145,19 +164,15 @@ public class ProgressActivity extends Activity {
         final Button button = (Button) findViewById(R.id.cancelPullButton);
         button.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
-                cancellPull();
+                cancelPull();
             }
         });
-
-        //the handler and the static activity is needed to show the dialog with the pull
-        // controller error.
-        handler = new Handler(Looper.getMainLooper());
-        progressActivity = this;
         intent = getIntent();
     }
 
-    private void cancellPull() {
-        mPullUseCase.cancell();
+    private void cancelPull() {
+        step(getBaseContext().getResources().getString(R.string.cancellingPull));
+        mPullUseCase.cancel();
     }
 
     @Override
@@ -185,34 +200,6 @@ public class ProgressActivity extends Activity {
         textView = (TextView) findViewById(R.id.pull_text);
     }
 
-    //@Subscribe
-    public void onProgressChange(final SyncProgressStatus syncProgressStatus) {
-        //bad event or activity not in foreground -> show nothing
-        if (syncProgressStatus == null || isOnPause) {
-            return;
-        }
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                if (syncProgressStatus.hasError()) {
-                    showException(syncProgressStatus.getException().getMessage() + " ");
-                    return;
-                }
-
-                //Step
-                if (syncProgressStatus.hasProgress()) {
-                    step(syncProgressStatus.getMessage() + " ");
-                    return;
-                }
-
-                //Finish
-                if (syncProgressStatus.isFinish()) {
-                    showAndMoveOn();
-                }
-            }
-        });
-    }
-
     /**
      * Launches a pull or push according to an intent extra
      */
@@ -232,25 +219,18 @@ public class ProgressActivity extends Activity {
     /**
      * Shows a dialog with the given message y move to login after showing error
      */
-    public static void showException(final String msg) {
+    public void showException(final String msg) {
         final boolean isAPush = isAPush();
         Log.d(TAG, msg + " ");
 
         PULL_ERROR = true;
         PULL_CANCEL = true;
-        handler.post(new Runnable() {
-            @Override
-            public void run() {
-                // Run your task here
-                new Handler(Looper.getMainLooper()).post(new Runnable() {
-                    @Override
-                    public void run() {
                         String dialogTitle = "", dialogMessage = "";
                         dialogTitle = getDialogTitle(isAPush());
                         if (msg != null) {
                             dialogMessage = msg;
                         }
-                        new AlertDialog.Builder(progressActivity)
+                        new AlertDialog.Builder(this)
                                 .setCancelable(false)
                                 .setTitle(dialogTitle)
                                 .setMessage(dialogMessage)
@@ -268,15 +248,11 @@ public class ProgressActivity extends Activity {
                                                 }
                                             }
                                         }).create().show();
-                    }
-                });
-            }
-        });
     }
 
-    private static void logout() {
+    private void logout() {
         Log.d(TAG, "Logging out...");
-        UserAccountRepository userAccountRepository = new UserAccountRepository(progressActivity);
+        UserAccountRepository userAccountRepository = new UserAccountRepository(this);
         LogoutUseCase logoutUseCase = new LogoutUseCase(userAccountRepository);
 
         logoutUseCase.execute(new LogoutUseCase.Callback() {
@@ -293,20 +269,12 @@ public class ProgressActivity extends Activity {
     }
 
     /**
-     * Prints the step in the progress bar
-     */
-    public static void step(final String msg) {
-        final int currentProgress = progressBar.getProgress();
-        progressBar.setProgress(currentProgress + 1);
-        textView.setText(msg);
-    }
-
-    /**
      * Shows a dialog to tell that pull is done and then moves into the dashboard.
      */
     private void showAndMoveOn() {
         if (PULL_ERROR) {
             PULL_ERROR = false;
+            finishAndGo(LoginActivity.class);
             return;
         }
         boolean isAPush = isAPush();
@@ -314,30 +282,31 @@ public class ProgressActivity extends Activity {
         //Annotate pull is done
         if (!isAPush) {
             //If is not active, we need restart the process
-            /*
-            if (!PULL_IS_ACTIVE) {
-                SdkController.unregister(this);
+            if (!mPullUseCase.isPullActive()) {
                 finishAndGo(LoginActivity.class);
                 return;
             } else {
+                //if is a pull and the process is finished, we annotate the first pull as true
                 annotateFirstPull(true);
             }
-            */
         }
 
         //Show final step -> done
         step(getString(R.string.progress_pull_done));
 
-        String title = getDialogTitle(isAPush);
+        String title = getDialogTitle(isAPush());
 
         final int msg = getDoneMessage();
 
         //Show message and go on -> pull or single push = dashboard | push before pull = start pull
         Intent intent = getIntent();
         //Not a pull -> is a Push
+        /**
+         * Reference required for testing purposes
+         */
         if (intent != null && (intent.getIntExtra(ProgressActivity.AFTER_ACTION,
                 ProgressActivity.DONT_SHOW_FEEDBACK) == ProgressActivity.SHOW_FEEDBACK)) {
-            this.alertDialog = new AlertDialog.Builder(this)
+            AlertDialog alertDialog = new AlertDialog.Builder(this)
                     .setCancelable(false)
                     .setTitle(title)
                     .setMessage(msg)
@@ -366,7 +335,7 @@ public class ProgressActivity extends Activity {
                             }).create();
             alertDialog.show();
         } else {
-            this.alertDialog = new AlertDialog.Builder(this)
+            AlertDialog alertDialog = new AlertDialog.Builder(this)
                     .setCancelable(false)
                     .setTitle(title)
                     .setMessage(msg)
@@ -410,30 +379,6 @@ public class ProgressActivity extends Activity {
         return R.string.dialog_push_success;
     }
 
-    private static void annotateFirstPull(boolean value) {
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(
-                PreferencesState.getInstance().getContext());
-        SharedPreferences.Editor editor = sharedPreferences.edit();
-        editor.putBoolean(
-                PreferencesState.getInstance().getContext().getString(R.string.pull_metadata),
-                value);
-        editor.commit();
-    }
-
-    /**
-     * Tells if a push is required
-     */
-    private static boolean isAPush() {
-        //A push before pull
-        if (pullAfterPushInProgress) {
-            return false;
-        }
-
-        //Not a pull -> is a Push
-        return (intent != null && intent.getIntExtra(TYPE_OF_ACTION, ACTION_PULL) != ACTION_PULL);
-    }
-
-
     /**
      * Tells is the intent requires a Pull after the push is done
      */
@@ -441,12 +386,6 @@ public class ProgressActivity extends Activity {
         Intent intent = getIntent();
         return (intent != null && intent.getIntExtra(TYPE_OF_ACTION, ACTION_PULL)
                 == ACTION_PUSH_BEFORE_PULL);
-    }
-
-    private static String getDialogTitle(boolean isAPush) {
-        int stringId =
-                isAPush ? R.string.dialog_title_push_response : R.string.dialog_title_pull_response;
-        return PreferencesState.getInstance().getContext().getString(stringId);
     }
 
     private void launchPull() {
@@ -459,8 +398,6 @@ public class ProgressActivity extends Activity {
                 EventExtended.format(month.getTime(), EventExtended.AMERICAN_DATE_FORMAT),
                 AppSettingsBuilder.isFullHierarchy(), AppSettingsBuilder.isDownloadOnlyLastEvents(),
                 PreferencesState.getInstance().getMaxEvents());
-        PopulateDB.wipeSDKData();
-        PopulateDB.wipeDatabase();
 
         mPullUseCase.execute(pullFilters, new PullUseCase.Callback() {
             @Override
@@ -470,16 +407,20 @@ public class ProgressActivity extends Activity {
 
             @Override
             public void onPullError() {
-                cancellPull(getBaseContext().getString(R.string
-                        .dialog_title_pull_response), getBaseContext().getString(R.string
+                showException(getBaseContext().getString(R.string
                         .dialog_pull_error));
             }
 
             @Override
+            public void onCancel() {
+                showException(getBaseContext().getString(R.string
+                        .pull_cancelled));
+            }
+
+            @Override
             public void onConversionError() {
-                cancellPull(getBaseContext().getString(R.string
-                        .dialog_title_pull_response), getBaseContext().getString(R.string
-                        .error_message));
+                showException(getBaseContext().getString(R.string
+                        .error_in_pull_conversion));
             }
 
             @Override
@@ -551,8 +492,9 @@ public class ProgressActivity extends Activity {
      * Find the surveys that are going to be pushed
      */
     private List<Survey> findSurveysToPush() {
+        List<Survey> surveys = new ArrayList<>();
         if (hasAPullAfterPush()) {
-            List<Survey> surveys = Survey.getAllUnsentUnplannedSurveys();
+            surveys = Survey.getAllUnsentUnplannedSurveys();
             for (int i = 0; i < surveys.size(); i++) {
                 if (surveys.get(i).getCompletionDate() == null) {
                     surveys.get(i).setCompleteSurveyState(Constants.PROGRESSACTIVITY_MODULE_KEY);
@@ -560,10 +502,6 @@ public class ProgressActivity extends Activity {
             }
             return surveys;
         }
-
-        List<Survey> surveys = new ArrayList<>();
-        //Fixme it is not used anymore?.
-        //surveys.add(Session.getSurveyByModule(module));
         return surveys;
     }
 
@@ -572,74 +510,14 @@ public class ProgressActivity extends Activity {
      *
      * @param targetActivityClass Given target activity class
      */
-    public static void finishAndGo(Class targetActivityClass) {
-        Intent targetActivityIntent = new Intent(progressActivity, targetActivityClass);
-        progressActivity.finish();
-        progressActivity.startActivity(targetActivityIntent);
+    public void finishAndGo(Class targetActivityClass) {
+        Intent targetActivityIntent = new Intent(this, targetActivityClass);
+        this.finish();
+        this.startActivity(targetActivityIntent);
     }
 
-    public static void cancellPull(final String title, final String errorMessage) {
-        PULL_ERROR = true;
-
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                // Run your task here
-                PULL_CANCEL = true;
-                new Handler(Looper.getMainLooper()).post(new Runnable() {
-                    @Override
-                    public void run() {
-                        String dialogTitle = "", dialogMessage = "";
-                        if (title != null) {
-                            dialogTitle = title;
-                        }
-                        if (errorMessage != null) {
-                            dialogMessage = errorMessage;
-                        }
-                        try {
-                            new AlertDialog.Builder(progressActivity)
-                                    .setCancelable(false)
-                                    .setTitle(dialogTitle)
-                                    .setMessage(dialogMessage)
-                                    .setNeutralButton(android.R.string.ok,
-                                            new DialogInterface.OnClickListener() {
-                                                public void onClick(DialogInterface arg0,
-                                                        int arg1) {
-                                                    Intent targetActivityIntent = new Intent(
-                                                            progressActivity, LoginActivity.class);
-                                                    targetActivityIntent.setFlags(
-                                                            Intent.FLAG_ACTIVITY_NEW_TASK);
-                                                    progressActivity.getApplicationContext()
-                                                            .startActivity(
-
-                                                                    targetActivityIntent);
-                                                    return;
-                                                }
-                                            }).create().show();
-                        } catch (Exception e) {
-                            Log.d(TAG, errorMessage);
-                        }
-
-                    }
-                });
-            }
-        }, 1000);
-    }
-
-    /**
-     * Checks if a dialog is being shown (for testing purposes)
-     */
-    public boolean isDialogShowing() {
-        boolean isShowing = this.alertDialog != null && this.alertDialog.isShowing();
-        Log.d(TAG, "isDialogShowing ->" + isShowing);
-        return isShowing;
-    }
-
-    public static void postFinish() {
-
-        ProgressActivity.annotateFirstPull(true);
-        finishAndGo(DashboardActivity.class);
-        //// FIXME: 28/11/2016 
-        //showAndMoveOn();
+    public void postFinish() {
+        annotateFirstPull(true);
+        showAndMoveOn();
     }
 }
