+++ app/src/main/java/org/eyeseetea/malariacare/domain/boundary/IPullControllerCallback.java
@@ -1,6 +1,6 @@
 package org.eyeseetea.malariacare.domain.boundary;
 
-import org.eyeseetea.malariacare.domain.entity.PullStep;
+import org.eyeseetea.malariacare.domain.usecase.pull.PullStep;
 
 public interface IPullControllerCallback {
     void onComplete();
@@ -8,4 +8,6 @@ public interface IPullControllerCallback {
     void onStep(PullStep step);
 
     void onError(Throwable throwable);
+
+    void onCancel();
 }
