--- app/src/main/java/org/eyeseetea/malariacare/domain/entity/PullStep.java
@@ -1,6 +1,0 @@
-package org.eyeseetea.malariacare.domain.entity;
-
-public enum PullStep {
-    PROGRAMS, EVENTS, PREPARING_PROGRAMS, PREPARING_ANSWERS, PREPARING_QUESTIONS,
-    PREPARING_RELATIONSHIPS, PREPARING_ORGANISATION_UNITS, VALIDATE_COMPOSITE_SCORES, PREPARING_SURVEYS
-}
